// types/person.ts
type Person = {
  biography: string | null;
  birthday: string | null;
  deathday: string | null;
  gender: number | null;
  id: number;
  imdb_id: string | null;
  known_for_department: string | null;
  name: string;
  nominations: number;
  place_of_birth: string | null;
  popularity: number | null;
  profile_path: string | null;
  tmdb_id: number | null;
  wins: number;
};

export default Person;